requirements.txt
streamlit
streamlit-player
kaggle
langchain
revChatGPT
deepl
streamlit-option-menu
pytest
python-dotenv
pyrebase4